week4_day1,3 과제 추가 제출
Day1
- 파일 이름 변경
Day3
- Mission1 Titanic 수정
- Mission2 제출(Exam scores for students at a public school)
